+++ pkg/managers/encoding.go
