+++ pkg/managers/managers_test.go
