+++ pkg/managers/shared_behavior.go
