+++ pkg/managers/shared_behavior_test.go
