+++ pkg/utils/encoder.go
