+++ pkg/utils/previews.go
